# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RUKİYE EKİNCİ Oyun Platformu API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get all posts (feed)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [all, photo, video, article]
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"

    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Wrong password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/games/{id}/approve:
    post:
      operationId: approveGame
      tags: [admin]
      summary: Approve a pending game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/games/{id}/reject:
    post:
      operationId: rejectGame
      tags: [admin]
      summary: Reject a pending game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: Game rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users:
    get:
      operationId: getAllUsers
      tags: [admin]
      summary: Get all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{id}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban or unban a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User banned/unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/posts:
    get:
      operationId: getAllPosts
      tags: [admin]
      summary: Get all posts for admin
      parameters:
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
  /admin/content/{type}/{id}:
    delete:
      operationId: deleteContent
      tags: [admin]
      summary: Delete any content (game, post, comment)
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [game, post, comment]
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/settings:
    get:
      operationId: getSettings
      tags: [admin]
      summary: Get platform settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"

    put:
      operationId: updateSettings
      tags: [admin]
      summary: Update platform settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettings"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        displayName:
          type: string
      required: [username, email, password]
    LoginRequest:
      type: object
      properties:
        emailOrUsername:
          type: string
        password:
          type: string
      required: [emailOrUsername, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        followersCount:
          type: integer
        followingCount:
          type: integer
        gamesCount:
          type: integer
        postsCount:
          type: integer
        createdAt:
          type: string
      required: [id, username, email, isAdmin, isBanned, followersCount, followingCount, gamesCount, postsCount, createdAt]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
          nullable: true
        mediaUrl:
          type: string
          nullable: true
        mediaType:
          type: string
          nullable: true
          enum: [photo, video, null]
        type:
          type: string
          enum: [photo, video, article, text]
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: string
          nullable: true
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        createdAt:
          type: string
      required: [id, type, authorId, authorName, likesCount, commentsCount, isLiked, createdAt]
    CreatePostRequest:
      type: object
      properties:
        content:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [photo, video]
        type:
          type: string
          enum: [photo, video, article, text]
      required: [type]
    PostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [posts, total, page, totalPages]
...
    AdminLoginRequest:
      type: object
      properties:
        password:
          type: string
      required: [password]
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required: [success, token]
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, totalPages]
    RejectRequest:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    PlatformSettings:
      type: object
      properties:
        siteName:
          type: string
        siteDescription:
          type: string
        allowRegistration:
          type: boolean
        requireGameApproval:
          type: boolean
        maxFileSize:
          type: integer
        featuredGameId:
          type: integer
          nullable: true
        maintenanceMode:
          type: boolean
      required: [siteName, siteDescription, allowRegistration, requireGameApproval, maxFileSize, maintenanceMode]
...