# lib/api-zod/src/generated/types/postListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RUKİYE EKİNCİ Oyun Platformu API
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from "./post";

export interface PostListResponse {
  posts: Post[];
  total: number;
  page: number;
  totalPages: number;
}