# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RUKİYE EKİNCİ Oyun Platformu API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email: string;
  displayName?: string;
  avatar?: string | null;
  bio?: string | null;
  isAdmin: boolean;
  isBanned: boolean;
  followersCount: number;
  followingCount: number;
  gamesCount: number;
  postsCount: number;
  createdAt: string;
}